requirements.txt
uvicorn
pydantic
fastapi
loguru
redis
tortoise-orm
aerich
aiomysql